requirements.txt
boto==2.48.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coverage==4.4.1
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.3
django-braces==1.11.0
django-map-widgets==0.1.9
django-model-utils==3.0.0
django-picklefield==1.0.0
django-registration==2.2
django-storages==1.6.3
factory-boy==2.8.1
Faker==0.7.17
funcsigs==1.0.2
geocoder==1.23.2
idna==2.5
ipaddress==1.0.18
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.1
python-google-places==1.4.0
pytz==2017.2
ratelim==0.1.6
requests==2.18.2
six==1.10.0
sorl-thumbnail==12.4a1
urllib3==1.22
